Move NullableJsonElementConverter to its own file
Also implements changes to `JsonElementConverter` from #311 in `NullableJsonElementConverter`
